Перевод проэкта с php

+++ e-vending/src/main/java/com/model/User.java
@@ -2,6 +2,7 @@ package com.model;
 
 import java.util.HashSet;
 import java.util.Set;
+
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -15,110 +16,110 @@ import javax.persistence.Table;
 @Table(name = "users")
 public class User {
 
-    @Id
-    @Column(name = "user_id", unique = true, nullable = false)
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+	@Id
+	@Column(name = "user_id", unique = true, nullable = false)
+	@GeneratedValue(strategy = GenerationType.AUTO)
+	private Long		id;
 
-    @Column(name = "root_user")
-    private Integer root_user;
+	@Column(name = "root_user")
+	private Integer		root_user;
 
-    @Column(name = "username")
-    private String username;
+	@Column(name = "username")
+	private String		username;
 
-    @Column(name = "fullname")
-    private String fullname;
+	@Column(name = "fullname")
+	private String		fullname;
 
-    @Column(name = "password")
-    private String password;
+	@Column(name = "password")
+	private String		password;
 
-    @Column(name = "enabled")
-    private boolean enabled;
+	@Column(name = "enabled")
+	private boolean		enabled;
 
-    @Column(name = "email")
-    private String email;
+	@Column(name = "email")
+	private String		email;
 
-    @ManyToMany(cascade = CascadeType.ALL, mappedBy = "users")
-    private Set<Role> userRole = new HashSet<>();
+	@ManyToMany(cascade = CascadeType.ALL, mappedBy = "users")
+	private Set<Role>	userRole	= new HashSet<>();
 
-    public User() {
-    }
+	public User() {
+	}
 
-    public User(String username, String password, boolean enabled) {
-        this.username = username;
-        this.password = password;
-        this.enabled = enabled;
-    }
+	public User(String username, String password, boolean enabled) {
+		this.username = username;
+		this.password = password;
+		this.enabled = enabled;
+	}
 
-    public User(String username, String password, boolean enabled, Set<Role> userRole) {
-        this.username = username;
-        this.password = password;
-        this.enabled = enabled;
-        this.userRole = userRole;
-    }
+	public User(String username, String password, boolean enabled, Set<Role> userRole) {
+		this.username = username;
+		this.password = password;
+		this.enabled = enabled;
+		this.userRole = userRole;
+	}
 
-    public String getUsername() {
-        return this.username;
-    }
+	public String getUsername() {
+		return this.username;
+	}
 
-    public void setUsername(String username) {
-        this.username = username;
-    }
+	public void setUsername(String username) {
+		this.username = username;
+	}
 
-    public String getPassword() {
-        return this.password;
-    }
+	public String getPassword() {
+		return this.password;
+	}
 
-    public void setPassword(String password) {
-        this.password = password;
-    }
+	public void setPassword(String password) {
+		this.password = password;
+	}
 
-    public boolean isEnabled() {
-        return this.enabled;
-    }
+	public boolean isEnabled() {
+		return this.enabled;
+	}
 
-    public void setEnabled(boolean enabled) {
-        this.enabled = enabled;
-    }
+	public void setEnabled(boolean enabled) {
+		this.enabled = enabled;
+	}
 
-    public Set<Role> getUserRole() {
-        return this.userRole;
-    }
+	public Set<Role> getUserRole() {
+		return this.userRole;
+	}
 
-    public void setUserRole(Set<Role> userRole) {
-        this.userRole = userRole;
-    }
+	public void setUserRole(Set<Role> userRole) {
+		this.userRole = userRole;
+	}
 
-    public Integer getId() {
-        return id;
-    }
+	public Long getId() {
+		return id;
+	}
 
-    public void setId(Integer id) {
-        this.id = id;
-    }
+	public void setId(Long id) {
+		this.id = id;
+	}
 
-    public String getEmail() {
-        return email;
-    }
+	public String getEmail() {
+		return email;
+	}
 
-    public void setEmail(String email) {
-        this.email = email;
-    }
+	public void setEmail(String email) {
+		this.email = email;
+	}
 
-    public String getFullname() {
-        return fullname;
-    }
+	public String getFullname() {
+		return fullname;
+	}
 
-    public void setFullname(String fullname) {
-        this.fullname = fullname;
-    }
+	public void setFullname(String fullname) {
+		this.fullname = fullname;
+	}
 
-    public void setRoot_user(Integer root_user) {
-        this.root_user = root_user;
-    }
+	public void setRoot_user(Integer root_user) {
+		this.root_user = root_user;
+	}
 
-    public Integer getRoot_user() {
-        return root_user;
-    }
+	public Integer getRoot_user() {
+		return root_user;
+	}
 
 }
